+++ crates/stackable-versioned/CHANGELOG.md
@@ -14,6 +14,7 @@ All notable changes to this project will be documented in this file.
 
 ### Changed
 
+- Remove duplicated code and unified struct/enum and field/variant code ([#820]).
 - Change from derive macro to attribute macro to be able to generate code
   _in place_ instead of _appending_ new code ([#793]).
 - Improve action chain generation ([#784]).
@@ -24,6 +25,7 @@ All notable changes to this project will be documented in this file.
 [#793]: https://github.com/stackabletech/operator-rs/pull/793
 [#804]: https://github.com/stackabletech/operator-rs/pull/804
 [#813]: https://github.com/stackabletech/operator-rs/pull/813
+[#820]: https://github.com/stackabletech/operator-rs/pull/820
 [#822]: https://github.com/stackabletech/operator-rs/pull/822
 
 ## [0.1.0] - 2024-05-08